+++ src/layouts/PlaceDetail/BottomBar/styled.ts
